+++ user-consumer-service/src/main/java/com/roncoo/eshop/model/InvestmentDetailsDO.java
@@ -7,11 +7,11 @@ import java.util.Date;
 
 @Data
 public class InvestmentDetailsDO {
-    private Integer id;
+    private Long id;
 
-    private Integer projectId;
+    private Long projectId;
 
-    private Integer investmenterId;
+    private Long investmenterId;
 
     private String projectName;
 
@@ -33,27 +33,27 @@ public class InvestmentDetailsDO {
 
     private Integer deleted;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
 
-    public Integer getInvestmenterId() {
+    public Long getInvestmenterId() {
         return investmenterId;
     }
 
-    public void setInvestmenterId(Integer investmenterId) {
+    public void setInvestmenterId(Long investmenterId) {
         this.investmenterId = investmenterId;
     }
